README.md
From the Glacier Point parking area, take the short, paved, wheelchair-accessible path to the exhilarating and hair-raising lookout point, 3,214' above Half Dome Village, on the floor of Yosemite Valley. The views are expansive and truly breathtaking. From this vantage point, visitors will be able to take in splendid views of waterfalls and the icnoic granite features like Half Dome.